requirements.txt
# Multi-Agent Debate System Dependencies

# LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0

# AWS Bedrock
boto3>=1.34.0
botocore>=1.34.0

# Search Tool
tavily-python>=0.3.0

# Environment Management
python-dotenv>=1.0.0

# Type Hints
typing-extensions>=4.9.0

# Optional: For better CLI experience
rich>=13.7.0
click>=8.1.0
